backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pillow==10.4.0
pypdf==4.3.1
img2pdf==0.5.1
pdf2docx==0.5.8
email-validator

